Another round of linting...
Move mathics.eval.distance to where it belongs
Some linting found during writing CosineDistance (more I hope to come though)
